fix developと同じ内容に修正

+++ back/src/test/java/mitei/mitei/political/balancesheet/manage/kanrensha/service/kanrensha/SearchPartnerAllByHistoryServiceTest.java
@@ -34,13 +34,13 @@ class SearchPartnerAllByHistoryServiceTest { // NOPMD
     private SearchPartnerAllByHistoryService searchPartnerAllByHistoryService;
 
     /** 関連者区分個人 */
-    private final static int PERSON = 1;
+    private final static short PERSON = 1;
 
     /** 関連者区分個人 */
-    private final static int CORP = 2;
+    private final static short CORP = 2;
 
     /** 関連者区分個人 */
-    private final static int POLI_ORG = 3;
+    private final static short POLI_ORG = 3;
 
     /** 010006,北海道 */
     @Test
